requirements.txt
astroid==2.3.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pycrypto==2.6.1
six==1.14.0
typed-ast==1.4.1
wrapt==1.11.2
